Added Patients_Invoices Report
Made Changes to the Main Menu so depending on user permissions they may not be able to access the Invoices Report. Invoices Report allows user to look up invoices based on first, last, name and or date range. The user can then delete an invoice or email parties that owe a balance by clicking on the email button.